Fix configuration page and form fields validation.
Configuration page has been broken because Form.findMatchingInput
function didn't work as expected. It was decided not to investigate
this issue but simply replace this function use with
'this.form.elements[required input].value'. Also at server side we
should decode values in order to support whitespaces in parameters
values.